worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = ""
authors = [
    {name = "Ricardo Alves da Silva",email = "ricardcpu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=12.0.0,<13.0.0)",
    "face-recognition (>=1.3.0,<2.0.0)",
    "face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models.git"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
